refactor: localize DB schema to Indonesian + admin fakultas v3.0.0
Changes:
- Align all GORM column tags in models.go with Indonesian schema
- Student, Achievement, Beasiswa, TiketAspirasi, BookingKonseling fixed
- TiketAspirasi: UUID primary key retained
- Create 04_localized_schema.sql (definitive public schema, 25+ tables)
- Refactor 02_admin_fakultas_schema.sql to v3.0.0 (all names in Indonesian)
- New columns: semester_sekarang, diperbarui_pada, catatan, diverifikasi_pada
- New columns: deskripsi/wajib (pkkmb_kegiatan), thumbnail/dilihat (berita)
- 10 new tables: periode_akademik, pengajuan_surat, program_mbkm,
  organisasi_mahasiswa, proposal_ormawa, proposal_fakultas,
  pkkmb_materi, pkkmb_tugas, pkkmb_kelulusan, pengumuman
- Add 11 new GORM structs in admin_fakultas.go
- Update FakAchievement, FakStudent to Indonesian column names
- Fix student_sync.go: CurrentSemester -> SemesterSekarang
- Fix achievement_handler.go + seed_prestasi.go: Tanggal -> Tahun
- Mark 13 English-table structs DEPRECATED
- Update README_SIAKAD.md with full schema docs and changelog

+++ backend/scripts/seeders/seed_reports.go
@@ -50,7 +50,7 @@ func SeedReports() {
 			NIM:             nim,
 			Name:            name,
 			MajorID:         major.ID,
-			EntryYear:       int16(year),
+			EntryYear:       year,
 			GPA:             2.0 + rand.Float64()*(4.0-2.0), // 2.0 to 4.0
 			Status:          statuses[rand.Intn(len(statuses))],
 			Gender:          genders[rand.Intn(len(genders))],
